fix: 테스트 진행 전 컨테이너 시작 설정 및 테스트 환경 통합 #7
