+++ src/test/java/org/songeun/petdongne_server/address/support/ElasticsearchIntegrationTestSupport.java
@@ -1,0 +1,42 @@
+package org.songeun.petdongne_server.address.support;
+
+import org.songeun.petdongne_server.address.infrastructure.elasticsearch.document.repository.AddressDocumentRepository;
+import org.songeun.petdongne_server.address.infrastructure.elasticsearch.index.AddressIndexRepository;
+import org.songeun.petdongne_server.address.infrastructure.elasticsearch.search.AddressSearchRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.elasticsearch.ElasticsearchContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+@SpringBootTest
+@ActiveProfiles("test")
+@Testcontainers
+public abstract class ElasticsearchIntegrationTestSupport {
+
+    @Container
+    static ElasticsearchContainer elasticsearchContainer =
+            new ElasticsearchContainer("docker.elastic.co/elasticsearch/elasticsearch:7.17.10")
+                    .withReuse(true);
+
+    @DynamicPropertySource
+    static void overrideProps(DynamicPropertyRegistry registry) {
+        if (!elasticsearchContainer.isRunning()) {
+            elasticsearchContainer.start();
+        }
+        registry.add("spring.elasticsearch.uris", elasticsearchContainer::getHttpHostAddress);
+    }
+
+    @Autowired
+    protected AddressSearchRepository searchRepository;
+
+    @Autowired
+    protected AddressDocumentRepository documentRepository;
+
+    @Autowired
+    protected AddressIndexRepository addressIndexRepository;
+
+}
